Setup for separate job module
